### 2022-12-13 #5010 @editorialbot
[PRE REVIEW]: MiNAA: Microbiome Network Alignment Algorithm
**Submitting author:** @crsl4 (<a href="http://orcid.org/0000-0002-9789-8915">Claudia Solis-Lemus</a>)
**Repository:** https://github.com/solislemuslab/minaa
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
