### 2022-12-13 #5010 @crsl4
Should we fix the invalid DOIs now or do we wait for more instructions?
