### 2022-12-17 4c2a957
fixed errors identified by JOSS editorial process
