+++ paper/paper.bib
@@ -1,115 +1,21 @@
-@inproceedings{kour2014real,
-  title={Real-time segmentation of on-line handwritten arabic script},
-  author={Kour, George and Saabne, Raid},
-  booktitle={Frontiers in Handwriting Recognition (ICFHR), 2014 14th International Conference on},
-  pages={417--422},
-  year={2014},
-  organization={IEEE}
-}
-
-@inproceedings{kour2014fast,
-  title={Fast classification of handwritten on-line Arabic characters},
-  author={Kour, George and Saabne, Raid},
-  booktitle={Soft Computing and Pattern Recognition (SoCPaR), 2014 6th International Conference of},
-  pages={312--318},
-  year={2014},
-  organization={IEEE}
-}
-
-@article{hadash2018estimate,
-  title={Estimate and Replace: A Novel Approach to Integrating Deep Neural Networks with Existing Applications},
-  author={Hadash, Guy and Kermany, Einat and Carmeli, Boaz and Lavi, Ofer and Kour, George and Jacovi, Alon},
-  journal={arXiv preprint arXiv:1804.09028},
-  year={2018}
-}
-
-
-@article{HyDe,
-	abstract = {{The analysis of hybridization and gene flow among closely related taxa is a common goal for researchers studying speciation and phylogeography. Many methods for hybridization detection use simple site pattern frequencies from observed genomic data and compare them to null models that predict an absence of gene flow. The theory underlying the detection of hybridization using these site pattern probabilities exploits the relationship between the coalescent process for gene trees within population trees and the process of mutation along the branches of the gene trees. For certain models, site patterns are predicted to occur in equal frequency (i.e., their difference is 0), producing a set of functions called phylogenetic invariants. In this article, we introduce HyDe, a software package for detecting hybridization using phylogenetic invariants arising under the coalescent model with hybridization. HyDe is written in Python and can be used interactively or through the command line using pre-packaged scripts. We demonstrate the use of HyDe on simulated data, as well as on two empirical data sets from the literature. We focus in particular on identifying individual hybrids within population samples and on distinguishing between hybrid speciation and gene flow. HyDe is freely available as an open source Python package under the GNU GPL v3 on both GitHub (https://github.com/pblischak/HyDe) and the Python Package Index (PyPI: https://pypi.python.org/pypi/phyde).}},
-	author = {Blischak, Paul D and Chifman, Julia and Wolfe, Andrea D and Kubatko, Laura S},
-	doi = {10.1093/sysbio/syy023},
-	eprint = {https://academic.oup.com/sysbio/article-pdf/67/5/821/25517390/syy023.pdf},
-	issn = {1063-5157},
-	journal = {Systematic Biology},
-	month = {03},
-	number = {5},
-	pages = {821-829},
-	title = {{HyDe: A Python Package for Genome-Scale Hybridization Detection}},
-	url = {https://doi.org/10.1093/sysbio/syy023},
-	volume = {67},
-	year = {2018},
-	bdsk-url-1 = {https://doi.org/10.1093/sysbio/syy023}}
-
-
-@misc{lev1athan,
-  doi = {10.48550/ARXIV.0910.4067},
-  
-  url = {https://arxiv.org/abs/0910.4067},
-  
[…]
