### 2022-12-17 #5010 @crsl4
@Kevin-Mattheus-Moerman Fixed DOIs and added more details to affiliations in [this commit](https://github.com/solislemuslab/minaa/commit/4c2a957224c8795d1117e1726f442c2547c54e2d).
