+++ paper/paper.bib
@@ -73,7 +73,7 @@
   number={13},
   pages={2182--2189},
   year={2015},
-  publisher={Oxford University Press}.
+  publisher={Oxford University Press},
   doi={10.1093/bioinformatics/btv130}
 }
 
