### 2022-12-17 #5010 @crsl4
Fixed silly error in paper.bib [here](https://github.com/solislemuslab/minaa/commit/35d88ab7abbfe1553cc3972c5eefc0306ea8e291).
