### 2023-01-10 #5010 @Kevin-Mattheus-Moerman
@crsl4 our board has determined this work may be in scope for JOSS. I have therefore removed the query-scope label. We will now proceed to seek a handling editor.
