+++ README.md
@@ -1,6 +1,6 @@
 # MiNAA: Microbiome Network Alignment Algorithm
 
-<img src="logo.png" style="float: right; max-width:40%;">
+<img src="logo.png" style="width:40%;" align=right>
 
 [![GitHub Releases](https://img.shields.io/github/v/release/solislemuslab/minaa?display_name=tag)](https://github.com/solislemuslab/minaa/releases) [![GitHub license](https://img.shields.io/github/license/solislemuslab/minaa)](https://github.com/solislemuslab/minaa/blob/main/LICENCE) [![GitHub Issues](https://img.shields.io/github/issues/solislemuslab/minaa)](https://github.com/solislemuslab/minaa/issues) ![ ](https://img.shields.io/github/languages/code-size/solislemuslab/minaa)
 
