+++ src/minaa.cpp
@@ -41,12 +41,16 @@ int main(int argc, char* argv[])
         auto bio_f = args[3];                      // biological data file
         auto alpha = std::stod(args[4]);           // GDV - edge weight balancer
         auto beta = std::stod(args[5]);            // topological - biological balancer
+        auto g_alias = args[6];                    // graph G alias
+        auto h_alias = args[7];                    // graph H alias
+        auto bio_alias = args[8];                  // biological data alias
 
         bool bio = (bio_f != "");                  // biological data file provided
 
         // Generate output names
-        auto g_name = FileIO::name_file(g_f);
-        auto h_name = FileIO::name_file(h_f);
+        auto g_name = FileIO::name_file(g_f, g_alias);
+        auto h_name = FileIO::name_file(h_f, h_alias);
+        auto bio_name = FileIO::name_file(bio_f, bio_alias);
         auto folder = FileIO::name_folder(g_name, h_name, datetime);
         log = folder + "log.txt";
 
@@ -61,9 +65,9 @@ int main(int argc, char* argv[])
         FileIO::out(log, "Y_M_D-H_M_S: " + datetime + "\n");
         FileIO::out(log, "\n");
         FileIO::out(log, "ARGUMENTS\n");
-        FileIO::out(log, "G File:   " + g_name + "\n");
-        FileIO::out(log, "H File:   " + h_name + "\n");
-        FileIO::out(log, "Bio File: " + bio_f + "\n");
+        FileIO::out(log, "G File:   " + g_name + ".csv\n");
+        FileIO::out(log, "H File:   " + h_name + ".csv\n");
+        if (bio) FileIO::out(log, "Bio File: " + bio_name + ".csv\n");
         FileIO::out(log, "Alpha:    " + Util::to_string(alpha, 3) + "\n");
         FileIO::out(log, "Beta:     " + Util::to_string(beta, 3) + "\n");
         FileIO::out(log, "\n");
@@ -71,13 +75,15 @@ int main(int argc, char* argv[])
         FileIO::out(log, "BEGINNING ALIGNMENT\n");
         auto s = std::chrono::high_resolution_clock::now();
 
-        // Read graph files into graph objects
+        // Read graph files into graph objects, write back to files
         FileIO::out(log, "Reading graph files............................");
         auto s00 = std::chrono::high_resolution_clock::now();
         auto g_graph = FileIO::file_to_graph(g_f);
         auto h_graph = FileIO::file_to_graph(h_f);
         auto g_labels = FileIO::parse_labels(g_f);
         auto h_labels = FileIO::parse_labels(h_f);
+        FileIO::graph_to_file(folder, g_name, g_labels, g_graph);
+        FileIO::graph_to_file(folder, h_name, h_labels, h_graph);
         auto f00 = std::chrono::high_resolution_clock::now();
         auto d00 = std::chrono::duration_cast<std::chrono::milliseconds>(f00-s00).count();
         FileIO::out(log, "done. (" + std::to_string(d00) + "ms)\n");
@@ -89,9 +95,9 @@ int main(int argc, char* argv[])
         auto h_gc_f = FileIO::graphcrunch_in(h_f, folder + h_name);
         auto g_gdvs = GraphCrunch::graphcrunch(g_gc_f);
         auto h_gdvs = GraphCrunch::graphcrunch(h_gc_f);
-        std::string rm1 = folder + g_name + ".csv";
+        std::string rm1 = folder + g_name + "_gc.csv";
         remove(rm1.c_str());
-        std::string rm2 = folder + h_name + ".csv";
+        std::string rm2 = folder + h_name + "_gc.csv";
         remove(rm2.c_str());
         auto f10 = std::chrono::high_resolution_clock::now();
[…]

[…]
