### 2023-06-03 c969828
make folder timestamp optional
