+++ src/minaa.cpp
@@ -44,15 +44,15 @@ int main(int argc, char* argv[])
         auto g_alias = args[6];                    // graph G alias
         auto h_alias = args[7];                    // graph H alias
         auto bio_alias = args[8];                  // biological data alias
-        auto passthrough = (args[9] == "1");       // passthrough?
-
-        bool bio = (bio_f != "");                  // biological data file provided?
+        auto do_passthrough = (args[9] == "1");    // do a passthrough of input files?
+        auto do_timestamp = (args[10] == "1");     // include a timestamp in the folder name?
+        auto do_bio = (bio_f != "");               // biological data file provided?
 
         // Generate output names
         auto g_name = FileIO::name_file(g_f, g_alias);
         auto h_name = FileIO::name_file(h_f, h_alias);
         auto bio_name = FileIO::name_file(bio_f, bio_alias);
-        auto folder = FileIO::name_folder(g_name, h_name, datetime);
+        auto folder = FileIO::name_folder(g_name, h_name, datetime, do_timestamp);
         log = folder + "log.txt";
 
         // Write log file
@@ -65,10 +65,10 @@ int main(int argc, char* argv[])
         FileIO::out(log, "\n");
         FileIO::out(log, "Y_M_D-H_M_S: " + datetime + "\n");
         FileIO::out(log, "\n");
-        FileIO::out(log, "ARGUMENTS\n");
+        FileIO::out(log, "INPUTS\n");
         FileIO::out(log, "G File:   " + g_name + ".csv\n");
         FileIO::out(log, "H File:   " + h_name + ".csv\n");
-        if (bio) FileIO::out(log, "Bio File: " + bio_name + ".csv\n"); // double .csv if no alias
+        if (do_bio) FileIO::out(log, "Bio File: " + bio_name + ".csv\n"); // double .csv if no alias
         FileIO::out(log, "Alpha:    " + Util::to_string(alpha, 3) + "\n");
         FileIO::out(log, "Beta:     " + Util::to_string(beta, 3) + "\n");
         FileIO::out(log, "\n");
@@ -87,9 +87,9 @@ int main(int argc, char* argv[])
         auto d00 = std::chrono::duration_cast<std::chrono::milliseconds>(f00-s00).count();
         FileIO::out(log, "done. (" + std::to_string(d00) + "ms)\n");
 
-        // Write graph objects back to files
-        if (passthrough)
+        if (do_passthrough)
         {
+            // Write graph objects back to files
             FileIO::out(log, "Writing graph files............................");
             auto s01 = std::chrono::high_resolution_clock::now();
             FileIO::graph_to_file(folder, g_name, g_labels, g_graph);
@@ -141,7 +141,7 @@ int main(int argc, char* argv[])
         
         std::vector<std::vector<double>> alignment;
 
-        if (bio) // incorporate biological data
+        if (do_bio) // incorporate biological data
         {
             // Parse and normalize the biological cost matrix
             FileIO::out(log, "Processing biological data.....................");
@@ -152,7 +152,7 @@ int main(int argc, char* argv[])
             auto d30 = std::chrono::duration_cast<std::chrono::milliseconds>(f30-s30).count();
             FileIO::out(log, "done. (" + std::to_string(d30) + "ms)\n");
 
-            if (passthrough)
+            if (do_passthrough)
             {
                 // Do a passthrough of the biological data file?
             }
