### 2023-06-17 bfaf305
update readme for new options
