+++ README.md
@@ -16,11 +16,13 @@ This program requires C++20 or higher.
 
 ### Unix
 
-`make`
+```bash
+make
+```
 
 ### Windows
 
-```cmd
+```bash
 mkdir obj
 make
 ```
@@ -29,41 +31,60 @@ In addition to C++20, Windows requires a special means to run the provided makef
 
 ## Usage
 
-This utility takes two to five command-line arguments: `./minaa.exe <G> <H> [-B=bio] [-a=alpha] [-b=beta]`
-
 ### Required Arguments (ordered)
 
-1. **G**: A network to align.
-2. **H**: A network to align.
+1. G; A network to align.
+2. H; A network to align.
 
 - Require:
   - The networks are represented by adjacency matrices in CSV format, with labels in both the first column and row.
   - The CSV delimiter must be one of {comma, semicolon, space, tab}, and will be detected automatically.
   - Any nonzero entry is considered an edge.
-  - |G| should be lesser or equal to |H|.
+  - |G| is lesser or equal to |H|.
 
 ### Optional Arguments (unordered)
 
-- **bio**: The path to the biological cost matrix file.
+#### Common
+
+- **-B=**: The path to the biological cost matrix file.
   - Require: CSV adjacency matrix where the first column is the labels of G, and first row is the labels of H.
   - Default: the algorithm will run using only topological calculations.
-- **alpha**: GDV-edge weight balancer
+- **-a=**: alpha; the GDV-edge weight balancer.
   - Require: a real number in range [0, 1].
   - Default: 1
-- **beta**: topological-biological cost matrix balancer.
+- **-b=**: beta; the topological-biological cost matrix balancer.
   - Require: a real number in range [0, 1].
   - Default: 1
+
+#### Uncommon
+
+- **-Galias=**: an alias for the G file.
+  - Require: a valid file name.
+  - Default: the G file keeps its original name.
+- **-Halias=**: an alias for the H file.
+  - Require: a valid file name.
+  - Default: the H file keeps its original name.
+- **-Balias=**: an alias for the B file.
+  - Require: a valid file name.
+  - Default: the B file keeps its original name.
+- **-p**: passthrough; whether or not to write the input files into the output folder.
+  - Require: none.
+  - Default: the files are not passed through to the output folder.
+- **-t**: Timestamp; the output folder's name includes the date and time of execution.
+  - Require: none.
+  - Default: the output folder's name does not include date and time.
 
 ### Outputs
 
-- **X-Y-T/**: (where "X", "Y" are the input networks, "T" is the date and time of execution) The folder containing the output files specified below.
-- **log.txt**: Record of the important details from the alignment.
-- **X_gdvs.csv**: (where "X" is the input network) The Graphlet Degree Vectors for network "X".
-- **top_costs.csv**: The topological cost matrix.
-- **bio_costs.csv**: The biologocal cost matrix (as inputed). Not created unless biological input is given.
-- **overall_costs.csv**: The combination of the topological and biological cost matrix. Not created unless biological input is given.
-- **alignment_list.csv**: A complete list of all aligned nodes, with rows in the format `g_node,h_node,similarity`, descending acording to similarity. The first row in this list is the total cost of the alignment, or the sum of (1 - similarity) for all aligned pairs.
-- **alignment_matrix.csv**: A matrix form of the same alignment, where the first column and row are the labels from the two input networks, respectively.
+- **G-H-T/**: (where G, H are the input networks, T is the date and time of execution) The folder containing the output files specified below.
+- **log.txt**: record of the important details from the alignment.
+- **G_gdvs.csv**: (where G is the input network) the Graphlet Degree Vectors for network G.
[…]
