### 2023-06-17 cb8c050
implement greekstamp
