### 2023-07-06 #5448 @waynebhayes
Hi, sorry I've been struggling a bit with Long COVID. I'll take a look at
the paper now.

On Tue, May 30, 2023 at 5:52 PM Gracielle Higino ***@***.***>
wrote:

> Hi @waynebhayes
> <https://urldefense.com/v3/__https://github.com/waynebhayes__;!!CzAuKJ42GuquVTTmVmPViYEvSg!OuSVd4GU4O7B5fukkqrcXB7GSGYUucDw9ZTykwbRXk6rgZ-Fp3cYNy89ERgwHaR4ZKZXLcOOfKcNsZKHgYQZqpU$>!
> Let me know if you're having trouble generating the checklist or adding
> your comments to this thread. =D
>
> @crsl4
> <https://urldefense.com/v3/__https://github.com/crsl4__;!!CzAuKJ42GuquVTTmVmPViYEvSg!OuSVd4GU4O7B5fukkqrcXB7GSGYUucDw9ZTykwbRXk6rgZ-Fp3cYNy89ERgwHaR4ZKZXLcOOfKcNsZKHUBAqd3Y$>
> just so you know, @Becheler
> <https://urldefense.com/v3/__https://github.com/Becheler__;!!CzAuKJ42GuquVTTmVmPViYEvSg!OuSVd4GU4O7B5fukkqrcXB7GSGYUucDw9ZTykwbRXk6rgZ-Fp3cYNy89ERgwHaR4ZKZXLcOOfKcNsZKHIpnWmh8$>
> has asked for an extension to complete their review for personal reasons
> and I've granted it. @Becheler
> <https://urldefense.com/v3/__https://github.com/Becheler__;!!CzAuKJ42GuquVTTmVmPViYEvSg!OuSVd4GU4O7B5fukkqrcXB7GSGYUucDw9ZTykwbRXk6rgZ-Fp3cYNy89ERgwHaR4ZKZXLcOOfKcNsZKHIpnWmh8$>
> please keep us up to date about your new intended timeline. [=
>
