### 2023-07-06 #5448 @waynebhayes
Um... it says "type @editorialbot generate my checklist"... where am I
supposed to type that?

I've read the paper and, from a scientific standpoint, it's woefully
lacking. It's leveraging ancient software for network alignment that's no
longer competitive; their algorithm offers nothing new over and above
existing ones; and they don't describe any *real* data their algorithm can
use. So I'm reluctant to spend much time on this. Furthermore... there
already exist many proven algorithms for network alignment, including my
own, github.com/waynebhayes/SANA, which even has a web interface.
