### 2023-07-07 #5448 @graciellehigino
Hi @waynebhayes

Thank you for having taken a look at the software. You should use the `@editorialbot` commands on this thread on GitHub, always as the first line of your comment. The bot will then generate a checklist just like the one that was generated for @Becheler to guide you on the things that matter to us in a review.

That being said, would you be willing to finish your review in the next 3 weeks?

Thank you!
