### 2023-07-08 #5448 @waynebhayes
Yes I could do it in 3 weeks.

On Fri, Jul 7, 2023 at 10:32 AM Gracielle Higino ***@***.***>
wrote:

> Hi @waynebhayes
> <https://urldefense.com/v3/__https://github.com/waynebhayes__;!!CzAuKJ42GuquVTTmVmPViYEvSg!IJE5F1OuKJQcdzCSjYzab2PTE4O044vFPptk8yIwQZT5P4VTkBmNQq_cIX8vN587CMMZ5jzghYlOZfqavz1ukUA$>
>
> Thank you for having taken a look at the software. You should use the
> @editorialbot commands on this thread on GitHub, always as the first line
> of your comment. The bot will then generate a checklist just like the one
> that was generated for @Becheler
> <https://urldefense.com/v3/__https://github.com/Becheler__;!!CzAuKJ42GuquVTTmVmPViYEvSg!IJE5F1OuKJQcdzCSjYzab2PTE4O044vFPptk8yIwQZT5P4VTkBmNQq_cIX8vN587CMMZ5jzghYlOZfqafURFfKw$>
> to guide you on the things that matter to us in a review.
>
> That being said, would you be willing to finish your review in the next 3
> weeks?
>
> Thank you!
>
