### 2023-07-08 #5448 @crsl4
Also, if I could also point out, most existing network alignment algorithms do not allow a user-specified biological similarity matrix. Is this an available option in SANA? We are also working on a web app, but it's taken a bit longer than expected, but in any case, we think that the inclusion of this matrix will provide flexibility to the algorithm.
