### 2023-07-14 #5448 @graciellehigino
@waynebhayes perfect! When you're ready, just visit the GitHub issue thread and copy-paste the command below:
`@editorialbot generate my checklist`

When you're done going through the checklist, please comment again saying that you're finished and adding any other comments you might have. Also, please address @crsl4's comment above.

Thank you!
