### 2023-07-14 #5448 @waynebhayes
Any algorithm that uses sequence similarity must allow some way to input
that similarity; it wouldn't be too hard to substitute any other sort of
similarity you want; about the only possible issue is normalizing, since
sequence similarity can be measured in many ways (identity, p-value, bit
score, etc) and it wouldn't necessarily be clear how to map the similarity
you WANT to use to the similarity that's expected by the algorithm. Still,
I think it would be possible with most algorithms.  And yes, SANA is one
such algorithm.

Also,  be careful using the word "matrix". You probably don't REALLY want
to use a matrix format, since most protein pairs will have a similarity of
zero, ie., the matrix would be very sparse and there's no point in storing
it that way or requiring users to upload it.

On Sat, Jul 8, 2023 at 9:26 AM Claudia Solis-Lemus ***@***.***>
wrote:

> Also, if I could also point out, most existing network alignment
> algorithms do not allow a user-specified biological similarity matrix. Is
> this an available option in SANA? We are also working on a web app, but
> it's taken a bit longer than expected, but in any case, we think that the
> inclusion of this matrix will provide flexibility to the algorithm.
>
