### 2023-07-14 #5448 @Becheler
> and it wouldn't necessarily be clear how to map the similarity you WANT to use to the similarity that's expected by the algorithm
In that case, it just means that the details of the similarity computation are not the responsibility of the algorithm, and can/must be abstracted out, for example with a callback defined by the user and invoked by the algorithm.

> be careful using the word "matrix". You probably don't REALLY want to use a matrix format, since most protein pairs will have a similarity of zero
I think it's fair use of the term as it helps biologists understanding the general idea of the object at hand, and software engineers will be able to general the concept easily. Again, if the details and efficiency of the matrix implementation (sparse, symetric etc) are not the responsibility of the algorithm, they can/should be abstracted out of said algorithm, for example by defining a callback that the algorithm can invoke e.g. `similarity(i,j)` without worrying of how this similiarity is  actually computed.
