### 2023-07-14 #5448 @crsl4
I could also add that by "biological similarity", we don't only mean "sequence similarity". Note that the networks here are microbiome networks. The input to create the networks (again, not responsibility of the current algorithm as the networks are taken as given) are usually relative abundances.

I also agree with @Becheler. There are sparse representations of matrices (e.g. [SparseArrays](https://docs.julialang.org/en/v1/stdlib/SparseArrays/)) which indeed denoted "matrices".
