### 2023-07-25 #5448 @waynebhayes
On Fri, Jul 14, 2023 at 2:06 AM Arnaud Becheler ***@***.***>
wrote:

> and it wouldn't necessarily be clear how to map the similarity you WANT to
>> use to the similarity that's expected by the algorithm
>
>

> In that case, it just means that the details of the similarity computation
> are not the responsibility of the algorithm,
>
> Correct.

> and can/must be abstracted out, for example with a callback defined by the
> user and invoked by the algorithm
>
> Not at all. You just need to describe what the algorithm expects, and the
user can provide their similarity scaled appropriately.

> be careful using the word "matrix". You probably don't REALLY want to use
>> a matrix format, since most protein pairs will have a similarity of zero
>
>

> I think it's fair use of the term as it helps biologists understanding the
> general idea of the object at hand, and software engineers will be able to
> general the concept easily.
>
> Fine.

> Again, if the details and efficiency of the matrix implementation (sparse,
> symetric etc) are not the responsibility of the algorithm, they can/should
> be abstracted out of said algorithm, for example by defining a callback
> that the algorithm can invoke e.g. similarity(i,j) without worrying of
> how this similiarity is actually computed.
>
>
I'm confused. "callback" doesn't solve the problem of how to scale the
similarity, it's just a different method for providing the data to the
algorithm. Using a "matrix" is fine as you first suggested; the "callback"
suggestion is completely a red herring unless I'm missing something.
