### 2023-07-26 #5448 @waynebhayes
On Fri, Jul 14, 2023 at 6:54 AM Claudia Solis-Lemus <
***@***.***> wrote:

> I could also add that by "biological similarity", we don't only mean
> "sequence similarity".
>
We've already established that. Many network alignment algorithms can
accept any sort of similarity, and as pointed out in another message, it's
the user's responsibility to provide the similarity scores. My point is
that the paper---unless there is a more recent version?---claims, "yet these
methods are restricted to one specific type of biological similarity (e.g.
sequence similarity in L-GRAAL)." *This is simply not true*. Not even
remotely. The algorithms take any type of similarity you want to provide.
You can provide sequence-based similarity, or Resnik scores based on Gene
Ontology annotations, or any of the several types of graphlet similarity,
or relative abundances as in your case. The similarity is just a number.
For alignment scoring purposes it doesn't matter what it represents.

The paper also says "In addition, unlike most existing network alignment
methods that are tailored for protein or gene interaction networks"... *this
statement is false**.* Network alignment algorithms align *networks*.
*PERIOD.* They can be social networks, PPI networks, gene networks,
microbiome networks, random networks.... it just so happens that the ones
you've looked at were *applied* to protein networks. They could just have
easily been applied to any other types of networks... *including* microbiome
ones.

> My primary concern with this work is that the authors don't seem to be at
all familiar with the existing literature on network alignment, or how to
use any of the existing algorithms. They don't define in their paper what
the similarity scores *mean*, and how they are scaled--for example, are
they logarithmic, or linear? Should they be in [0,1], or can they be
arbitrary? Are negative numbers allowed?

And there are no comparisons to existing algorithms at all--and as I said
above, you *cannot* claim that other algorithms "only do sequence". That's
false. So I'm still not convinced the algorithm actually does anything
novel that's not already done by existing algorithms.

> Note that the networks here are microbiome networks. The input to create
> the networks (again, not responsibility of the current algorithm as the
> networks are taken as given) are usually relative abundances.
>
Fine. So try aligning them with an existing algorithm. Try SANA. It also
has a web interface at http://sana.ics.uci.edu. Or just send me the
networks you'd like to align. I'd be happy to try SANA on them on your
behalf.

I also agree with @Becheler
> <https://urldefense.com/v3/__https://github.com/Becheler__;!!CzAuKJ42GuquVTTmVmPViYEvSg!J0BG36PasVe5bAlJ38wl-OYZHtXESGJ5bLm6FD3928AklMzRO6W8Ssh_7iCN8orIlrIPO_5MoepnZDzHsjfJguA$>.
> There are sparse representations of matrices (e.g. SparseArrays
> <https://urldefense.com/v3/__https://docs.julialang.org/en/v1/stdlib/SparseArrays/__;!!CzAuKJ42GuquVTTmVmPViYEvSg!J0BG36PasVe5bAlJ38wl-OYZHtXESGJ5bLm6FD3928AklMzRO6W8Ssh_7iCN8orIlrIPO_5MoepnZDzHXyScyJA$>)
> which indeed denoted "matrices".
>
Agreed.
