### 2023-07-26 #5448 @Becheler
> Not at all. You just need to describe what the algorithm expects, and the
user can provide their similarity scaled appropriately.

> I'm confused. "callback" doesn't solve the problem of how to scale the
> similarity, it's just a different method for providing the data to the
> algorithm. Using a "matrix" is fine as you first suggested; the "callback"
> suggestion is completely a red herring unless I'm missing something.

Our two perspectives are neither exclusive nor contradictory. I think @waynebhayes is missing here several fundamentals of C++ software design, specifically concepts related to the decoupling of data structures and algorithms. When orthogonal aspects of an algorithm begin to be user context dependent rather than developer context dependent, then a popular idiom is to encapsulate these behaviors into a policy-class that implements the required interface. By writing a custom implementation of a given policy, the user of a policy-based library can inject custom behaviors unforeseen by the library implementor. From my perspective, all aspects mentioned by @waynebhayes (scaling, matrix implementation details such as sparse vs dense, meaning and assumptions of the similarity index etc) are all orthogonal to the alignment algorithm (this specific point @waynebhayes seems to understand). If the policy interface can be reduced to a simple callable operator (what I believe it does: taking a and b points and returning the similarity measured between them - that could be of any return type), then it boils down to requiring the algorithm to take an arbitrary callable (possibly fulfilling some requirements that could be checked with C++20 concepts). I recommend @waynebhayes getting familiar with the Open/Closed principle and reading  _Modern C++ Design: Generic Programming and Design Patterns Applied_ by Andrei Alexandrescu, specifically the chapter on Policy-Based Design.

> So I'm still not convinced the algorithm actually does anything novel that's not already done by existing algorithms.

I respectfully disagree. The paper the authors are submitting to is JOSS: unlike many other journals, JOSS focuses on the actual implementation of a software resource. As a Software Engineer in Genetics, I am more than happy to see a diverse crew of small code repositories, with a carefully tested implementation of small components, dedicated to a specific purpose. In the case of software design and code reuse, less is more, and I would personally rather work with the MINAA code base than the more imposing SANA that tightly entangles orthogonal concepts such as Graph, Alignment and SparseMatrix.

On another note, I am quite shocked by the insistance of @waynebhayes to compete and refer to their own resource. This is quite unethical reviewing and added to the unpleasant tone of the review, clearly lacks of humility. Contrary to what they say, it is actually surprisingly difficult in Software Engineering to find implementations that "do nothing new" and it is quite naive to believe a competing resource "does it all": most of the time, their idiosyncrasies would complement each other. I would encourage trying to be curious and understand what are the specificities of MINAA's implementation and hypothesis, why using SANA would not have been helpful to the authors in this context (as I said, SANA's tightly coupled class design may be the beginning of an answer), and how a better state of the art could be reached through cooperation and complementation.
Respectfully,
