### 2023-07-26 #5448 @waynebhayes
I'm sorry, I was under the impression that I was asked to review this
article as an expert in *network alignment* from the standpoint of whether
it constitutes a novel contribution to the *science* of network alignment.
It is my firm opinion that, within that context, it most definitely is *not* a
novel contribution in that it does nothing new that existing algorithms do
not do, does not compare itself to existing algorithms, does not
demonstrate that it can be applied to *real* networks of any
significant size, does not describe how a user interested in the *scientific
outcome* should use the algorithm... among probably other reasons.

In the above context, it is not at all unethical to suggest that a proposed
algorithm compare itself to existing algorithms, in order to demonstrate
novelty. This is a cornerstone of science. I did not *exclusively* suggest
SANA as an algorithm to compare against; I suggested it in this context
because, so far as I am aware, it's the only one (other than the one in the
current paper) that offers an easy-to-use Web interface. I was also
interested in the *scientific* outcome of aligning microbiome networks...
because I am acutely interested in *any* application of network alignment
to new data. My offer of applying SANA to the networks in the paper was
offered out of a genuine feeling of scientific curiosity, not out of any
intent to compete. (And my offer stands.)

If, as you seem to suggest, the goal of this paper and this algorithm is to
demonstrate software engineering concepts in a toy implementation applied
to a toy problem in the context, for example, of a software engineering
class, then I'm 100% OK with that.... but I don't *care* about that
application. If the intent of this paper and this algorithm is to claim
that it does something new that other alignment algorithms do not.... then
my answer, as an expert in the *scientific* field of *network alignment*...
is no. It does not.

I'm done. Thanks for reading.
