### 2023-07-26 #5448 @Becheler
I believe you can be an expert in a field and still exercise your expertise as a reviewer in a way that is mindful and considerate.

Research software is not purely about « demonstrating software engineering concepts » nor is it purely about the « scientific output ». It is about both, simply. And there is nothing about « toyish » in discussing about code design and reuse, as decades of open software libraries speak for themselves.

The question I ask regularly myself is « can I find a simple, testable, open source C++  implementation of this feature and reuse it for my own software science purposes ?». Applied to the context of this article, the answer is « no, there is no such existing codebase» and this makes the present software relevant to my needs. I also consider I could increment and build upon MINAA more easily than SANA, and this is food for thoughts.

I agree that in an ideal world we would have reviewed a flawless code with a flawless design, with memory and runtime benchmark comparison with X existing algorithms and Y data structures, with a complete documentation and an impressive real world data application, and why not a website and a GUI and a build system and a packaging for vcpkg and Conan and deb and also support for most of the compiler versions out there … but is it really reasonable? Such resources take time to build and elaborate upon, as we all know.

Science community has much to benefit from such little code projects, that should be encouraged and improved rather than frowned upon.
