### 2023-07-28 #5448 @crsl4
If I could also add, we are submitting to JOSS because MiNAA is a useful software product. JOSS does not require comparisons to other algorithms as part of the manuscript.

From the JOSS website:

JOSS publishes articles about research software. This definition includes software that: solves complex modeling problems in a scientific context (physics, mathematics, biology, medicine, social science, neuroscience, engineering); supports the functioning of research instruments or the execution of research experiments; extracts knowledge from large data sets; offers a mathematical library; or similar.

JOSS submissions must:
- Be open source (i.e., have an [OSI-approved license](https://opensource.org/licenses/category)).
- Have an obvious research application.
- Be feature-complete (no half-baked solutions) and be designed for maintainable extension (not one-off modifications).
- Minor 'utility' packages, including 'thin' API clients, and single-function packages are not acceptable.

We are not producing an new algorithm that outperforms every existing network alignment available (otherwise, we would be submitting as a research manuscript, not a software note). We produced a useful software product that is modular and easily extendable (as @Becheler points out).
I should also point out that even though networks are networks, they don't all have similar characteristics. For example, networks under different applications have different degree distributions or sparsity levels. We cannot assume that methods that perform well under one type of network will automatically perform well on others. While we have not made the extensive study of limitations of MiNAA (or other algorithms like SANA) under microbiome networks (again because this is a software note, not a research paper), we want MiNAA to be that easy-to-use tool that we can continuously extend based on our (and our users') data characteristics that are _specific_ to microbiome networks. There are many tools out there that have been developed and tested on protein interaction networks. Do they perform well on microbiome networks? Maybe, we don't know.
We do not care if MiNAA performs well on protein interaction networks or social networks. We want to provide a tool that is specific to microbiome networks and their inherent characteristics that distinguish them from other biological or social networks.

Perhaps some of the confusion here is that @waynebhayes is evaluating MiNAA as if it were a research paper, when this is a software note with different expectations, as described in the JOSS website. We are happy to change the language if @waynebhayes has specific suggestions.
