### 2023-08-30 #5448 @graciellehigino
Hi all!

After reading all these comments, considering that the software is working correctly and it _does_ fill a gap, I am happy to **recommend it for publication**.

@crsl4 has properly responded to all the comments from the reviewers, even when some of them were not in scope, which showed the strength of the publication.

I am really sorry that this review history shows some bad attitudes and that the authors had to read and deal with those comments, but I believe that the good thing about open peer review is to leave these marks so they never happen again.

On the other hand, this thread also shows some great attitudes and comments, and similarly, they are registered here so everyone can see how good researchers you are. Thank you so much for your constructive conversations, @Becheler and @crsl4

@Kevin-Mattheus-Moerman as this is a very particular situation, I would like to have you involved in the acceptance process as the EiC. If JOSS strongly believes we should start over or get a third reviewer, we can do so, but I do not see the need for that.
