### 2023-10-18 #5448 @graciellehigino
After talking with the EiC, we've decided that we do need a second reviewer to complete the process. We will work hard to find someone as soon as possible, and we appreciate your patience as we go through this together.
