### 2023-11-02 #5448 @Kevin-Mattheus-Moerman
@openjournals/dev :wave: I would like to trigger the search for similar JOSS papers, is that possible?
