### 2023-11-02 #5448 @Kevin-Mattheus-Moerman
@crsl4 could you please extend the description of your software in your README? At the moment is has a single sentence (with a typo) i.e. `MiNAA aligns two networks based their topologies and biologies.` which should likely read `MiNAA aligns two networks based on their topologies and biologies.`.
It would be great if you had a bit of a paragraph, perhaps using more plain language, so it would be clear to those seeing this README/project of the first time, what the software is about.
