### 2023-11-02 #5448 @Kevin-Mattheus-Moerman
@antagomir @microsud @grabear @angrybee @torognes @rdenise would you be interested in reviewing this submission on microbiome network allignment for [JOSS (the Journal of Open Source Software](https://joss.theoj.org)?

JOSS reviews focus on testing [the software](https://github.com/solislemuslab/minaa), and reviewing a very [short paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05448/joss.05448/10.21105.joss.05448.pdf).

Reviewing for JOSS is an easy/smooth process, and takes place right here on GitHub. This here is a review issue, and above you can see that one reviewer has completed their review already and has gone through the checklist guiding reviewers through the smooth process.

If you are interested please let us know by commenting here.
