### 2023-11-02 #5448 @xuanxu
> openjournals/dev 👋 I would like to trigger the search for similar JOSS papers, is that possible?

The `similar papers` search is only active on pre-review issues (by now), because it is more useful when reviewers are not assigned yet. The easiest way to trigger it now would be to use the `generate pdf`command in the [pre-review](https://github.com/openjournals/joss-reviews/issues/5010) of this submission.
