### 2023-11-17 #5448 @arfon
@Kevin-Mattheus-Moerman @graciellehigino – I ran the 'similar papers' task over here for you: https://github.com/openjournals/joss-reviews/issues/5010#issuecomment-1816084719
