### 2023-11-22 #5448 @Kevin-Mattheus-Moerman
@salix-d folks can most certainly volunteer. Our documentation for reviewers lives here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. If you can confirm you have no conflict of interest I can consider assigning you as reviewer if you like.
