### 2023-11-22 #5448 @salix-d
I can confirm I have no conflict of interest 🙂
I had done another JOSS review in May and was interested in this one too!
I was sorry to see it didn't work out with the other reviewer 🙁
