### 2023-11-26 #5448 @salix-d
Hi! Got stucked at the 'Installation instructions' point today;  see https://github.com/solislemuslab/minaa/issues/6.
