### 2023-11-26 4a500e3
clarify compilation instructions
