### 2023-11-27 solislemuslab/minaa#6 @reednel
Thank you for the detailed description @salix-d! the README has been updated to reflect the need for specifically g++, and with a link to the MinGW download from SourceForge. Let me know if you have any further suggestions.
