### 2023-11-27 solislemuslab/minaa#7 @salix-d
Output name error on Windows
So running the example command: `./minaa.exe example/G.csv example/H.csv`
I got this error:
```
ERROR:
 Unable to create output folder alignments\example/G-example/H-2023_11_26-18_56_14\
```
The problem is [at line 39 of file_io.cpp](https://github.com/solislemuslab/minaa/blob/4a500e3195133ffb24ed7267dc901bc03b3cd941/src/file_io.cpp#L39):
```
auto si = file.find_last_of("\\") + 1;
```
For windows you're looking for the last backslash but your example code uses slash.

If I run `./minaa.exe example\\G.csv example\\H.csv` instead it works, but to avoid this you could simply check for the last backslash or slash (`"\\|/"`) instead and then you could use the same command regardless of OS.

```
    /**
     * Returns the truncated name of the file.
     *
     * @param file The file to truncate.
     *
     * @return The truncated name of the file.
     */
    std::string name_file(std::string file)
    {
        auto si = file.find_last_of("\\|/") + 1;
        auto ei = file.find_last_of(".");
        auto file_name = file.substr(si, ei - si);
        return file_name;
    }
```
Rebuilt with this code ^ and got no error.
