### 2023-11-27 #5448 @salix-d
I also hit a bigger snag trying to make my c++ work on VScode but that was VScode issues ^^' (paths not being updated properly in the config json files type of thing)

Now that I got _that_ working, I did hit another small bump 😅
The file_io.cpp script is a bit too specific on the slashes allowed in files' name: see https://github.com/solislemuslab/minaa/issues/7

On one hand, I'm a bit annoyed at Windows for always having more issues.
On the other, I'm glad your software gets tested on it otherwise they wouldn't have been noticed. 😄
