### 2023-11-27 #5448 @salix-d
In that case, you might want to know that Windows has no problem running .sh files, so you didn't really need to do the .bat one.

### Re: the code
I did managed to run the examples and the simulations with no more problems 😄
I also tried the simulation with a different alpha value (0.6) to see if it would give similar results.
![heatmap_a06_web](https://github.com/openjournals/joss-reviews/assets/31168746/a0d82757-a48b-4c6d-a17e-6fa70a0f9f23)

### Re: the documentation
In the README file, in the example section, the code doesn't use the names of the file in the example folder.
I understand using more generic variable names as 'network0' and 'network1', but I would probably mention that there's a readme file and example data in the example folder, so if they want to try the examples they should use that code.

The 2nd example uses a 'bio_costs.csv' but there's no equivalent provided in the example folder, so we can't really test that functionality. Since being able to include measures of biological similarity seemed to be an important aspect of this tool, I would have expected an example of it. It's also not examplified in the paper.

### Re: the paper
I found one spelling error on line 102  "for the same networs"; should be "networks".
I noticed that the legend of figure 2 isn't labeled or described in the figure's text.
