### 2023-11-28 #5448 @reednel
In the latest push
- "networs" changed to "networks" in `paper/paper.md:76`.
- Added description of the figure 2 legend in `paper/paper.md:78`.
- Added a sentence pointing to `example/` in the Example section of the README.
- Revised the `name_file()` function of `src/file_io.cpp` according to Salix's suggestions.

The point about our absense of an example using biological data warrants discussion with @crsl4.

I'm keeping `simulations/align.bat` around for now because on my own Windows machine, I am not able to run `simulations/align.sh`, and from what I can find online, Windows cannot run bash scripts in the ordinary command line.
