### 2023-11-28 #5448 @Becheler
> I also hit a bigger snag trying to make my c++ work on VScode but that was VScode issues ^^' (paths not being updated properly in the config json files type of thing)
>
> Now that I got _that_ working, I did hit another small bump 😅 The file_io.cpp script is a bit too specific on the slashes allowed in files' name: see [solislemuslab/minaa#7](https://github.com/solislemuslab/minaa/issues/7)
>
> On one hand, I'm a bit annoyed at Windows for always having more issues. On the other, I'm glad your software gets tested on it otherwise they wouldn't have been noticed. 😄

Just a quick fact (not even a suggestion, it's beyond the scope of this article), but Visual Studio Code has a Devcontainer extension that allows you to connect to Docker (you just have to open Docker) and quickly sets up whatever image you defined in your `.devcontainer/Dockerfile` file and `.devcontainer/devcontainer.json`. Super easy to setup, so the user litterally just have to click on "open in container", and then click on whatever task you defined in your `.vscode/tasks.json` file (for example a "Compile and Run Tests" task). Very cool feature ;)
