### 2023-11-29 #5448 @reednel
> The point about our absense of an example using biological data warrants discussion with @crsl4.

Following up on this: Claudia, Rosa, and I are working out a method to synthesize the biological component. The tentative idea is a new R script that constructs a phylognentic tree for the nodes in the networks, and produces a distance matrix from there. We'll keep you updated as we go.
