+++ src/file_io.cpp
@@ -13,15 +13,15 @@
 namespace FileIO
 {
     /**
-     * Returns whether or not the given string is to a readable file.
+     * Returns whether or not the given string is a path to a readable file.
      * 
-     * @param file The file to check.
+     * @param filepath The filepath to check.
      * 
      * @return True if the file is readable, false otherwise.
      */
-    bool is_accessible(std::string file)
+    bool is_accessible(std::string filepath)
     {
-        std::ifstream fin(file);
+        std::ifstream fin(filepath);
         return fin.good();
     }
 
@@ -41,24 +41,24 @@ namespace FileIO
      * Detect the delimiter used for a CSV file.
      * Detectable delimiters are comma, semicolon, space, and tab.
      * 
-     * @param file The csv file to detect the delimiter of.
+     * @param filepath Path to the csv file to detect the delimiter of.
      * 
      * @return the delimiter used in the file.
      * 
      * @throws std::runtime_error If the file could not be opened.
      */
-    char detect_delimiter(std::string file)
+    char detect_delimiter(std::string filepath)
     {
         // Open the given file, read the first line
         std::ifstream fin;
         fin.exceptions(std::ofstream::badbit);
         try 
         {
-            fin.open(file);
+            fin.open(filepath);
         }
         catch (const std::ifstream::failure& e)
         {
-            throw std::runtime_error("Unable to open file " + file);
+            throw std::runtime_error("Unable to open file " + filepath);
         }
 
         std::string line;
@@ -96,13 +96,14 @@ namespace FileIO
     /**
      * Returns the truncated name of the file.
      * 
-     * @param file The file to truncate.
+     * @param filepath The filepath to truncate.
+     * @param given_name The name to give the file, if nonempty.
      * 
-     * @return The truncated name of the file.
+     * @return given_name if nonempty, else the truncated name of the file.
      */
-    std::string name_file(std::string file, std::string given_name)
+    std::string name_file(std::string filepath, std::string given_name)
     {
-        if (file == "")
+        if (filepath == "")
         {
             return "";
         }
@@ -111,9 +112,9 @@ namespace FileIO
             return given_name;
         }
 
-        auto start_index = file.find_last_of("/|\\") + 1;
-        auto end_index = file.find_last_of(".");
-        auto file_name = file.substr(start_index, end_index - start_index);
+        auto start_index = filepath.find_last_of("/|\\") + 1;
+        auto end_index = filepath.find_last_of(".");
+        auto file_name = filepath.substr(start_index, end_index - start_index);
         return file_name;
     }
 
@@ -122,7 +123,12 @@ namespace FileIO
      * 
      * @param g_name The name of the first graph file.
      * @param h_name The name of the second graph file.
-     * @param time The time of the run.
+     * @param datetime The date and time of the run.
+     * @param do_timestamp Whether or not to include the date and time in the folder name.
+     * @param do_greekstamp Whether or not to include the greek parameters in the folder name.
+     * @param alpha The alpha value used in the run.
+     * @param beta The beta value used in the run.
+     * @param do_bio Whether or not to include the biological input file in the folder name.
      * 
      * @return The name of the folder to store the output files in.
      * 
@@ -226,11 +232,11 @@ namespace FileIO
     /**
      * Outputs the given string to the given file and std::cout.
      * 
-     * @param file The file to output to.
+     * @param filepath The file to output to.
      * 
      * @throws std::runtime_error If the file could not be opened.
      */
-    void out(std::string file, std::string str)
+    void out(std::string filepath, std::string str)
     {
         std::cout << str << std::flush;
 
[…]
