+++ src/file_io.cpp
@@ -119,101 +119,106 @@ namespace FileIO
     }
 
     /**
-     * Names the folder to store the output files in, based on the input files and time.
+     * Names the directory to store the output files in, based on the input files and time.
      * 
+     * @param base_path The directory in which the named directory will be created.
      * @param g_name The name of the first graph file.
      * @param h_name The name of the second graph file.
      * @param datetime The date and time of the run.
-     * @param do_timestamp Whether or not to include the date and time in the folder name.
-     * @param do_greekstamp Whether or not to include the greek parameters in the folder name.
+     * @param do_timestamp Whether or not to include the date and time in the directory name.
+     * @param do_greekstamp Whether or not to include the greek parameters in the directory name.
      * @param alpha The alpha value used in the run.
      * @param beta The beta value used in the run.
-     * @param do_bio Whether or not to include the biological input file in the folder name.
+     * @param do_bio Whether or not a biological input was included.
      * 
-     * @return The name of the folder to store the output files in.
+     * @return The name of the directory to store the output files in.
      * 
-     * @throws std::runtime_error If a necessary folder/file could not be created.
+     * @throws std::runtime_error If a necessary directory/file could not be created.
      */
-    std::string name_folder(std::string g_name, std::string h_name, std::string datetime, bool do_timestamp, bool do_greekstamp, std::string alpha, std::string beta, bool do_bio)
+    std::string name_directory(std::string base_path, std::string g_name, std::string h_name, std::string datetime, bool do_timestamp, bool do_greekstamp, std::string alpha, std::string beta, bool do_bio)
     {
-        std::string folder = "";
+        std::string path = "";
 
         #ifdef _WIN32 // Windows
         {
-            const std::string OUTPUT_FOLDER = "alignments\\";
-
-            // Make directory "alignments\" if it doesn't exist
-            if (mkdir(OUTPUT_FOLDER.c_str()) == -1)
+            // Make base_path directory if it doesn't exist
+            if (mkdir(base_path.c_str()) == -1)
             {
                 if(errno != EEXIST)
                 {
-                    throw std::runtime_error("Unable to create directory " + OUTPUT_FOLDER);
+                    throw std::runtime_error("Unable to create directory " + base_path);
+                }
+            }
+
+            path = base_path + "\\" + g_name + "-" + h_name;
+
+            if (do_greekstamp)
+            {
+                path += "-a" + alpha;
+                if (do_bio)
+                {
+                    path += "-b" + beta;
                 }
             }
 
             if (do_timestamp)
             {
-                folder = OUTPUT_FOLDER + g_name + "-" + h_name + "-" + datetime + "\\";
+                path += "-" + datetime;
             } 
-            else
-            {
-                folder = OUTPUT_FOLDER + g_name + "-" + h_name + "\\";
-            }
+            
+            path += "\\";
             
             // Make directory if it doesn't exist
-            folder = OUTPUT_FOLDER + g_name + "-" + h_name + "-" + datetime + "\\";
-            if (mkdir(folder.c_str()) == -1)
+            if (mkdir(path.c_str()) == -1)
             {
                 if(errno != EEXIST)
                 {
-                    throw std::runtime_error("Unable to create output folder " + folder);
+                    throw std::runtime_error("Unable to create output directory " + path);
                 }
             }
         }
         #else // Unix
         {
-            const std::string OUTPUT_FOLDER = "alignments/";
-
-            // Make directory "alignments/" if it doesn't exist
[…]
