+++ src/minaa.cpp
@@ -45,16 +45,24 @@ int main(int argc, char* argv[])
         auto h_alias = args[7];                    // graph H alias
         auto bio_alias = args[8];                  // biological data alias
         auto do_passthrough = (args[9] == "1");    // do a passthrough of input files?
-        auto do_timestamp = (args[10] == "1");     // include a timestamp in the folder name?
-        auto do_greekstamp = (args[11] == "1");    // include a greekstamp in the folder name?
+        auto do_timestamp = (args[10] == "1");     // include a timestamp in the directory name?
+        auto do_greekstamp = (args[11] == "1");    // include a greekstamp in the directory name?
         auto do_bio = (bio_file != "");            // biological data file provided?
+
+        const auto BASE_PATH = "alignments";
+        const auto LOG_FILENAME = "log.txt";
+        const auto TOP_COSTS_FILENAME = "topological_costs.csv";
+        const auto BIO_COSTS_FILENAME = "biological_costs.csv";
+        const auto OVERALL_COSTS_FILENAME = "overall_costs.csv";
+        const auto ALIGNMENT_MATRIX_FILENAME = "alignment_matrix.csv";
+        const auto ALIGNMENT_LIST_FILENAME = "alignment_list.csv";
 
         // Generate output names
         auto g_name = FileIO::name_file(g_file, g_alias);
         auto h_name = FileIO::name_file(h_file, h_alias);
         auto bio_name = FileIO::name_file(bio_file, bio_alias);
-        auto folder = FileIO::name_folder(g_name, h_name, datetime, do_timestamp, do_greekstamp, args[4], args[5], do_bio);
-        log = folder + "log.txt";
+        auto directory = FileIO::name_directory(BASE_PATH, g_name, h_name, datetime, do_timestamp, do_greekstamp, args[4], args[5], do_bio);
+        log = directory + LOG_FILENAME;
 
         // Write log file
         FileIO::out(log, "Executing:   ");
@@ -93,8 +101,8 @@ int main(int argc, char* argv[])
             // Write graph objects back to files
             FileIO::out(log, "Writing graph files............................");
             auto s01 = std::chrono::high_resolution_clock::now();
-            FileIO::graph_to_file(folder, g_name, g_labels, g_graph);
-            FileIO::graph_to_file(folder, h_name, h_labels, h_graph);
+            FileIO::graph_to_file(directory + g_name + ".csv", g_labels, g_graph);
+            FileIO::graph_to_file(directory + h_name + ".csv", h_labels, h_graph);
             auto f01 = std::chrono::high_resolution_clock::now();
             auto d01 = std::chrono::duration_cast<std::chrono::milliseconds>(f01-s01).count();
             FileIO::out(log, "done. (" + std::to_string(d01) + "ms)\n");
@@ -103,13 +111,13 @@ int main(int argc, char* argv[])
         // Calculate the GDVs for G and H, remove temp files
         FileIO::out(log, "Calculating GDVs...............................");
         auto s10 = std::chrono::high_resolution_clock::now();
-        auto g_gc_file = FileIO::graphcrunch_in(g_file, folder + g_name);
-        auto h_gc_file = FileIO::graphcrunch_in(h_file, folder + h_name);
+        auto g_gc_file = FileIO::graphcrunch_in(g_file, directory + g_name);
+        auto h_gc_file = FileIO::graphcrunch_in(h_file, directory + h_name);
         auto g_gdvs = GraphCrunch::graphcrunch(g_gc_file);
         auto h_gdvs = GraphCrunch::graphcrunch(h_gc_file);
-        std::string rm1 = folder + g_name + "_gc.csv";
[…]
