+++ include/file_io.h
@@ -7,7 +7,7 @@ namespace FileIO
     bool isValidFileName(std::string);
     char detect_delimiter(std::string);
     std::string name_file(std::string, std::string);
-    std::string name_folder(std::string, std::string, std::string, bool, bool, std::string, std::string, bool);
+    std::string name_directory(std::string, std::string, std::string, std::string, bool, bool, std::string, std::string, bool);
     void out(std::string, std::string);
     void err(std::string, std::string);
 
@@ -17,11 +17,11 @@ namespace FileIO
     std::vector<std::string> parse_labels(std::string);
     std::vector<std::vector<double>> file_to_cost(std::string);
 
-    std::string graph_to_file(std::string, std::string, std::vector<std::string>, std::vector<std::vector<unsigned>>);
-    std::string gdvs_to_file(std::string, std::string, std::vector<std::string>, std::vector<std::vector<unsigned>>);
-    std::string cost_to_file(std::string, std::string, std::vector<std::string>, std::vector<std::string>, std::vector<std::vector<double>>);
-    std::string alignment_to_matrix_file(std::string, std::vector<std::string>, std::vector<std::string>, std::vector<std::vector<double>>);
-    std::string alignment_to_list_file(std::string, std::vector<std::string>, std::vector<std::string>, std::vector<std::vector<double>>);
+    void graph_to_file(std::string, std::vector<std::string>, std::vector<std::vector<unsigned>>);
+    void gdvs_to_file(std::string, std::vector<std::string>, std::vector<std::vector<unsigned>>);
+    void cost_to_file(std::string, std::vector<std::string>, std::vector<std::string>, std::vector<std::vector<double>>);
+    void alignment_to_matrix_file(std::string, std::vector<std::string>, std::vector<std::string>, std::vector<std::vector<double>>);
+    void alignment_to_list_file(std::string, std::vector<std::string>, std::vector<std::string>, std::vector<std::vector<double>>);
 }
 
 #endif
