### 2024-01-12 #5448 @reednel
We're going to get real data to construct a full example that includes the use of a meaningful biological matrix in the alignment. We can't say exactly when it will be ready, as we're depending on a couple others to get us the data and assemble it into something usable. But it's a priority for us so we can push this paper through. Once we get this example in and add it to the documentation, I believe we will have satisfied the remaining boxes on Salix's checklist :)
