### 2024-01-16 e7fb972
implement similarity to cost option
