### 2024-01-16 a31103b
implement similarity threshold
