### 2024-01-17 1e87730
overhaul examples
