### 2024-01-17 c4fbde9
overhaul examples
