+++ example/g-h-a0.6/log.txt
@@ -1,0 +1,19 @@
+Executing:   ./minaa.exe example/g.csv example/h.csv -a=0.6 -g 
+Y_M_D-H_M_S: 2024_01_16-22_02_40
+
+INPUTS
+G File:               g.csv
+H File:               h.csv
+Alpha:                0.600
+Beta:                 1.000
+Similarity threshold: 0.0000
+
+BEGINNING ALIGNMENT
+Reading graph files............................done. (9ms)
+Calculating GDVs...............................done. (19ms)
+Writing GDVs to files..........................done. (4ms)
+Calculating the topological cost matrix........done. (16ms)
+Writing the topological cost matrix to file....done. (5ms)
+Aligning the graphs............................done. (48ms)
+Writing the alignment to file..................done. (1ms)
+ALIGNMENT COMPLETED (106ms)
