+++ examples/README.md
@@ -1,6 +1,6 @@
 # Examples
 
-The purpose of this folder is to provide worked example of MiNAA's inputs, execution, and outputs.
+The purpose of this directory is to provide worked example of MiNAA's inputs, execution, and outputs.
 
 `g.csv` is the adjacency matrix for the network G.
 `h.csv` is the adjacency matrix for the network H.
@@ -10,7 +10,7 @@ Once MiNAA is successfully compiled, the examples below can be run from this pro
 
 ## Example 1
 
-`./minaa.exe example/g.csv example/h.csv -a=0.6 -g`
+`./minaa.exe examples/g.csv examples/h.csv -a=0.6 -g`
 
 Output to: `g-h-a0.6/`
 
@@ -18,7 +18,7 @@ Here we align network **g** with network **h** using no biological data. `-a=0.6
 
 ## Example 2
 
-`./minaa.exe example/g.csv example/h.csv -B=example/bio.csv -b=0.85 -st=0.5 -s`
+`./minaa.exe examples/g.csv examples/h.csv -B=examples/bio.csv -b=0.85 -st=0.5 -s`
 
 Output to: `g-h/`
 
@@ -26,7 +26,7 @@ Here we align network **g** with network **h** using topological information and
 
 ## Example 3
 
-`./minaa.exe example/g.csv example/h.csv -Galias=nonsmoker -Halias=smoker -p -t`
+`./minaa.exe examples/g.csv examples/h.csv -Galias=nonsmoker -Halias=smoker -p -t`
 
 Output to: `nonsmoker-smoker-2024_01_16-22_05_34/`
 
