+++ examples/g-h/log.txt
@@ -1,4 +1,4 @@
-Executing:   ./minaa.exe example/g.csv example/h.csv -B=example/bio.csv -b=0.85 -st=0.5 -s 
+Executing:   ./minaa.exe examples/g.csv examples/h.csv -B=examples/bio.csv -b=0.85 -st=0.5 -s 
 Y_M_D-H_M_S: 2024_01_16-22_00_30
 
 INPUTS
