+++ src/graphcrunch.cpp
@@ -32,81 +32,64 @@
 #include <string.h>
 #include <string>
 #include <assert.h>
-#include <map>        /* STL ordered dictionary class */
+#include <map> /* STL ordered dictionary class */
 #include <vector>
 
 namespace GraphCrunch
 {
-    //#define PATH_MAX 256 //Oleksii
-    
+    // #define PATH_MAX 256 //Oleksii
+
     /* These lookup tables were generated by analyzing sample graphs representing
-    * each graphlet, and then adjusting the table sizes for perfect hashing.
-    */
+     * each graphlet, and then adjusting the table sizes for perfect hashing.
+     */
 
     /* Lookup table for graphlets. Calculated by:
-    * [(sum of (degree % 4 for node/neighbors))/2][edge count/2]-4
-    * Special handling needed for X23 and X25
-    */
-    const char gtable [][8] =
-    {{-1,-1,10,-1,-1,8},{-1,11,-1,-1,15,14,12},{17,19,-1,16,18,20},
-    {-1,-1,23,24,-1,21},{-1,-1,26,25},{-1,-1,-1,-1,-1,-1,27},{28}};
+     * [(sum of (degree % 4 for node/neighbors))/2][edge count/2]-4
+     * Special handling needed for X23 and X25
+     */
+    const char gtable[][8] =
+        {{-1, -1, 10, -1, -1, 8}, {-1, 11, -1, -1, 15, 14, 12}, {17, 19, -1, 16, 18, 20}, {-1, -1, 23, 24, -1, 21}, {-1, -1, 26, 25}, {-1, -1, -1, -1, -1, -1, 27}, {28}};
 
     /* Lookup table for nodes by [graphlet][sum of degrees of node/neighbors] */
-    const char ntable [][21] =
-    {{-1,-1,-1,0,1},{-1,-1,-1,-1,-1,-1,2},{-1,-1,-1,3,-1,4},{-1,-1,-1,-1,5,-1,6},
-    {-1,-1,-1,-1,-1,-1,7},{-1,-1,-1,-1,8,-1,-1,9,10},
-    {-1,-1,-1,-1,-1,-1,-1,-1,11,-1,12},{-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,13},
-    {-1,-1,-1,14,-1,15,16},{-1,-1,-1,17,18,-1,19,20},{-1,-1,-1,-1,-1,21,-1,-1,22},
-    {-1,-1,-1,-1,23,-1,-1,-1,24,25},{-1,-1,-1,26,-1,-1,27,28,-1,29},
-    {-1,-1,-1,-1,-1,30,-1,-1,31,-1,32},{-1,-1,-1,-1,-1,-1,33},
-    {-1,-1,-1,-1,34,-1,35,36,37},{-1,-1,-1,-1,-1,38,-1,-1,-1,39,-1,40,41},
-    {-1,-1,-1,-1,-1,-1,-1,-1,42,-1,-1,-1,43},{-1,-1,-1,-1,44,-1,-1,-1,45,-1,46,47},
-    {-1,-1,-1,-1,-1,-1,-1,-1,48,49},{-1,-1,-1,-1,-1,-1,-1,50,51,-1,52},
-    {-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,53,-1,-1,-1,54},
-    {-1,-1,-1,-1,-1,55,-1,-1,-1,-1,-1,-1,-1,56,57},
[…]
