+++ src/minaa.cpp
@@ -17,7 +17,7 @@
 /*
  * Main function
  */
-int main(int argc, char* argv[])
+int main(int argc, char *argv[])
 {
     std::string log;
     try
@@ -30,26 +30,26 @@ int main(int argc, char* argv[])
         {
             args = Util::parse_args(argc, argv);
         }
-        catch (std::invalid_argument& e)
+        catch (std::invalid_argument &e)
         {
             std::cerr << "ERROR: " << e.what() << "\nPROGRAM TERMINATING" << std::endl;
             return 1;
         }
 
-        auto g_file = args[1];                              // graph G file
-        auto h_file = args[2];                              // graph H file
-        auto bio_file = args[3];                            // biological data file
-        auto alpha = std::stod(args[4]);                    // GDV - edge weight balancer
-        auto beta = std::stod(args[5]);                     // topological - biological balancer
-        auto similarity_threshold = std::stod(args[6]);     // similarity threshold above which alignments report
-        auto g_alias = args[7];                             // graph G alias
-        auto h_alias = args[8];                             // graph H alias
-        auto bio_alias = args[9];                           // biological data alias
-        auto do_passthrough = (args[10] == "1");            // do a passthrough of input files?
-        auto do_timestamp = (args[11] == "1");              // include a timestamp in the directory name?
-        auto do_greekstamp = (args[12] == "1");             // include a greekstamp in the directory name?
-        auto do_similarity_conversion = (args[13] == "1");  // convert biological similarity to costs?
-        auto do_bio = (bio_file != "");                     // biological data file provided?
+        auto g_file = args[1];                             // graph G file
+        auto h_file = args[2];                             // graph H file
+        auto bio_file = args[3];                           // biological data file
+        auto alpha = std::stod(args[4]);                   // GDV - edge weight balancer
+        auto beta = std::stod(args[5]);                    // topological - biological balancer
+        auto similarity_threshold = std::stod(args[6]);    // similarity threshold above which alignments report
+        auto g_alias = args[7];                            // graph G alias
+        auto h_alias = args[8];                            // graph H alias
+        auto bio_alias = args[9];                          // biological data alias
+        auto do_passthrough = (args[10] == "1");           // do a passthrough of input files?
+        auto do_timestamp = (args[11] == "1");             // include a timestamp in the directory name?
+        auto do_greekstamp = (args[12] == "1");            // include a greekstamp in the directory name?
+        auto do_similarity_conversion = (args[13] == "1"); // convert biological similarity to costs?
+        auto do_bio = (bio_file != "");                    // biological data file provided?
 
         const auto BASE_PATH = "alignments";
         const auto LOG_FILENAME = "log.txt";
@@ -98,7 +98,7 @@ int main(int argc, char* argv[])
         auto g_labels = FileIO::parse_labels(g_file);
         auto h_labels = FileIO::parse_labels(h_file);
         auto f00 = std::chrono::high_resolution_clock::now();
-        auto d00 = std::chrono::duration_cast<std::chrono::milliseconds>(f00-s00).count();
[…]
