+++ include/hungarian.h
@@ -5,5 +5,5 @@ namespace Hungarian
 {
     std::vector<std::vector<double>> hungarian(std::vector<std::vector<double>>);
 }
-    
+
 #endif
