### 2024-01-18 #5448 @reednel
We've pushed a reproducible example that uses real biological data, and a couple others. All can be found in the [examples folder](https://github.com/solislemuslab/minaa/tree/main/examples). At the same time we've pulled in additional updates we've been developing. For the full list of changes, see my notes on the [PR](https://github.com/solislemuslab/minaa/pull/13). To be clear, core functionality is the same, everything described in the manuscript is still accurate, and any previously run commands and inputs would yeild the same output. All new features are purely additional and optional. @salix-d and @Becheler, these features are documented in the README and examples of their usage are in the same examples folder if you'd like to look them over. Let me know of any questions or concerns. Thank you!
