### 2024-02-06 #5448 @graciellehigino
@crsl4 and @reednel could you go through [this checklist ](https://github.com/openjournals/joss-reviews/issues/5448#issuecomment-1928739036)with me before I recommend it for publication?
