### 2024-02-06 #5448 @reednel
Yes we'll get on that, thanks @graciellehigino. For box 2, the latest release with all changes from the review is [1.1.0](https://github.com/solislemuslab/minaa/releases/tag/v1.1.0).
