### 2024-02-06 #5448 @reednel
@graciellehigino on Zenodo I have the option to set the DOI or create a new one. Should I generate one or use one from the bot's list above, or somehwere else?
