### 2024-02-12 #5448 @reednel
Zenodo
- https://zenodo.org/records/10625654
- 10.5281/zenodo.10625654
Figshare
- https://figshare.com/articles/software/MiNAA_Microbiome_Network_Alignment_Algorithm/25153106
- 10.6084/m9.figshare.25153106

@graciellehigino All checkboxes on our side can be ticked. Let us know if anything needs revision.
