### 2024-03-19 #5448 @graciellehigino
> The paper's PDF and metadata files generation produced some warnings that could prevent the final paper from being published. Please fix them before the end of the review process.
>
> ```
> citation kurtz2015sparse not found
> ```

@crsl4 @reednel oops, I let it slide this reference! Could you fix it? As soon as you do, @Kevin-Mattheus-Moerman will be able to proceed with acceptance [=
