+++ paper/paper.bib
@@ -54,6 +54,18 @@
   doi={10.1002/nav.20053}
 }
 
+@article{kurtz2015sparse,
+  title={Sparse and Compositionally Robust Inference of Microbial Ecological Networks},
+  author={Kurtz, Zachary D. and M\{"u}ller, Christian L. and Miraldi, Emily R. and Littman, Dan R. and Blaser, Martin J. and Bonneau, Richard A.},
+  journal={PLOS Computational Biology},
+  volume={11},
+  number={5},
+  pages={1--25},
+  year={2015},
+  publisher={Public Library of Science},
+  doi={10.1371/journal.pcbi.1004226},
+}
+
 @article{ma2020review,
   title={A review of protein--protein interaction network alignment: From pathway comparison to global alignment},
   author={Ma, Cheng-Yu and Liao, Chung-Shou},
@@ -112,13 +124,13 @@
   doi={10.1101/2020.07.15.195248}
 }
 
-@article{Pilgrim1995,
-author = {Pilgrim, Robert},
-year = {1995},
-month = {08},
-journal = {Lecture notes},
-title = {Tutorial on Implementation of Munkres' Assignment Algorithm},
-doi = {10.13140/RG.2.1.3572.3287}
+@article{pilgrim1995,
+  author={Pilgrim, Robert},
+  year={1995},
+  month={08},
+  journal={Lecture notes},
+  title={Tutorial on Implementation of Munkres' Assignment Algorithm},
+  doi={10.13140/RG.2.1.3572.3287}
 }
 
 @article{prvzulj2007biological,
