### 2024-03-20 a554e09
remove fancy char
