+++ paper/paper.bib
@@ -56,7 +56,7 @@
 
 @article{kurtz2015sparse,
   title={Sparse and Compositionally Robust Inference of Microbial Ecological Networks},
-  author={Kurtz, Zachary D. and M\{"u}ller, Christian L. and Miraldi, Emily R. and Littman, Dan R. and Blaser, Martin J. and Bonneau, Richard A.},
+  author={Kurtz, Zachary D. and Muller, Christian L. and Miraldi, Emily R. and Littman, Dan R. and Blaser, Martin J. and Bonneau, Richard A.},
   journal={PLOS Computational Biology},
   volume={11},
   number={5},
