### 2024-04-05 #5448 @crsl4
@Kevin-Mattheus-Moerman I don't know if you need our confirmation, but the proof looks good.
