### 2024-04-05 #5448 @reednel
@Kevin-Mattheus-Moerman I created the Zenodo and Figshare archives, and did set the version to 1.1.0. See Zenodo here: https://zenodo.org/records/10625654. Or if you're referring to something else, can you point me to that?
