### 2024-04-05 #5448 @Kevin-Mattheus-Moerman
@reednel I meant that you should include the `V` in `V1.1.0`. Currently over at Zenodo it just says 1.1.0 but it should match your GitHub release tag which includes the `V`. Confusingly at the Zenodo archive it will end up saying `Version V1.1.0`.
You should be able to manually edit the version tag

Hope this helps clarify things.
