### 2024-04-05 #5448 @reednel
Understood, thanks for clarifying. Zenodo now reads `v1.1.0` to match the Github tag.
